+++ spec/app/app/serializers/panko/ar/user_serializer.rb
@@ -2,7 +2,7 @@ module Panko
   module Ar
     class UserSerializer < BaseSerializer
       typelize_from ::User
-      attributes :id, :username
+      attributes :id, :username, :skills, :settings, :metadata
 
       has_one :invitor, serializer: UserSerializer